requirements.txt
# --- Core ML stack ---
numpy<2
pandas>=2.2.2
Pillow>=10.4.0
opencv-python==4.8.1.78
matplotlib>=3.9.2
rasterio==1.4.3
tqdm>=4.66.5
scikit-learn>=1.5.2

# --- Experiment tracking / backend ---
mlflow>=2.5.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
pyyaml>=6.0.1
python-dotenv>=1.0.1
minio>=7.2.0

# --- Monitoring ---
prometheus-client>=0.20.0
